gamma_backend/requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

coverage
django
django-cors-headers
django-extensions
django-jenkins
djangorestframework
djangorestframework-jwt
geopy
gunicorn
pillow
psycopg2-binary
scipy
